+++ src/AceTimeClock.h
@@ -17,16 +17,8 @@
 #ifndef ACE_TIME_CLOCK_ACE_TIME_CLOCK_H
 #define ACE_TIME_CLOCK_ACE_TIME_CLOCK_H
 
-// Blacklist boards using new Arduino API due to incompatibilities. This
-// currently includes all megaAVR boards and SAMD21 boards using arduino::samd
-// >= 1.8.10. Boards using arduino:samd <= 1.8.9 or SparkFun:samd are fine.
-#if defined(ARDUINO_ARCH_MEGAAVR)
-#error MegaAVR not supported, https://github.com/bxparks/AceTime/issues/44
-
-#elif defined(ARDUINO_ARCH_SAMD) && defined(ARDUINO_API_VERSION)
-#error SAMD21 with arduino:samd >= 1.8.10 not supported, https://github.com/bxparks/AceTime/issues/45
-
-#elif defined(ARDUINO_API_VERSION)
+// Blacklist boards using new Arduino API due to incompatibilities.
+#if defined(ARDUINO_API_VERSION)
 #error Platforms using ArduinoCore-API not supported
 #endif
 
@@ -34,6 +26,7 @@
 #include "ace_time/clock/NtpClock.h"
 #include "ace_time/clock/DS3231Clock.h"
 #include "ace_time/clock/UnixClock.h"
+#include "ace_time/clock/EspSntpClock.h"
 #include "ace_time/clock/SystemClock.h"
 #include "ace_time/clock/SystemClockLoop.h"
 #include "ace_time/clock/SystemClockCoroutine.h"